refactor: changed type of variable to match with type of Java HashMap

+++ src/main/java/me/gfumagali/btgpedidos/repository/model/documents/ClientOrders.java
@@ -14,7 +14,7 @@ public class ClientOrders {
     @Id
     private Long clientCode;
 
-    private Long ordersQuantity;
+    private Integer ordersQuantity;
 
     private HashMap<Long, Order> orders;
 }
